move ValidatedTextField to CommonUiComponents
